+++ backend/src/main/java/com/sjiwon/anotherart/purchase/domain/repository/PurchaseRepository.java
@@ -1,5 +1,6 @@
-package com.sjiwon.anotherart.purchase.domain;
+package com.sjiwon.anotherart.purchase.domain.repository;
 
+import com.sjiwon.anotherart.purchase.domain.model.Purchase;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PurchaseRepository extends JpaRepository<Purchase, Long> {
