+++ backend/src/main/java/com/sjiwon/anotherart/purchase/presentation/PurchaseApiController.java
@@ -1,7 +1,7 @@
-package com.sjiwon.anotherart.purchase.controller;
+package com.sjiwon.anotherart.purchase.presentation;
 
 import com.sjiwon.anotherart.global.resolver.ExtractPayload;
-import com.sjiwon.anotherart.purchase.service.PurchaseService;
+import com.sjiwon.anotherart.purchase.application.usecase.PurchaseService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
