--- backend/src/test/java/com/sjiwon/anotherart/common/ServiceTest.java
@@ -1,50 +1,0 @@
-package com.sjiwon.anotherart.common;
-
-import com.sjiwon.anotherart.art.domain.repository.ArtRepository;
-import com.sjiwon.anotherart.auction.domain.repository.AuctionRepository;
-import com.sjiwon.anotherart.common.config.DatabaseCleanerEachCallbackExtension;
-import com.sjiwon.anotherart.common.config.MySqlTestContainersExtension;
-import com.sjiwon.anotherart.common.config.RedisTestContainersExtension;
-import com.sjiwon.anotherart.favorite.domain.repository.FavoriteRepository;
-import com.sjiwon.anotherart.member.domain.repository.MemberRepository;
-import com.sjiwon.anotherart.purchase.domain.PurchaseRepository;
-import com.sjiwon.anotherart.token.domain.repository.TokenRepository;
-import com.sjiwon.anotherart.token.utils.JwtTokenProvider;
-import jakarta.persistence.EntityManager;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.transaction.annotation.Transactional;
-
-@SpringBootTest
-@Transactional
-@ExtendWith({
-        DatabaseCleanerEachCallbackExtension.class,
-        MySqlTestContainersExtension.class,
-        RedisTestContainersExtension.class
-})
-public abstract class ServiceTest {
-    @Autowired
-    protected EntityManager em;
-
-    @Autowired
-    protected JwtTokenProvider jwtTokenProvider;
-
-    @Autowired
-    protected MemberRepository memberRepository;
-
-    @Autowired
-    protected TokenRepository tokenRepository;
-
-    @Autowired
-    protected ArtRepository artRepository;
-
-    @Autowired
-    protected FavoriteRepository favoriteRepository;
-
-    @Autowired
-    protected AuctionRepository auctionRepository;
-
-    @Autowired
-    protected PurchaseRepository purchaseRepository;
-}
